part 4 - refactor tree (no grid) items https://online.sbis.ru/opendoc.html?guid=d26eb342-c3ff-4892-84af-27819a43196a
