+++ Controls/_tree/display/TreeCollection.ts
@@ -56,7 +56,6 @@ export default class TreeCollection<
 
     isLastItem(item: CollectionItem): boolean {
         // TODO Сделать возможным делать последний NodeFooter last, если он содержит данные
-        //  Сейчас прямо в шаблоне проверяется как shouldDisplayVisibleFooter(content)
         //  Как можно проверить из кода - не ясно
         const enumerator = this._getUtilityEnumerator();
