+++ Controls/_tree/display/TreeNodeFooterItem.ts
@@ -1,87 +1,29 @@
 import { TemplateFunction } from 'UI/Base';
-import { Model } from 'Types/entity';
-import TreeItem from './TreeItem';
+import TreeNodeExtraItem from './TreeNodeExtraItem';
 
-export default class TreeNodeFooterItem extends TreeItem<null> {
-    readonly '[Controls/tree:TreeNodeFooterItem]': boolean;
-    readonly Markable: boolean = false;
-    readonly Fadable: boolean = false;
-    readonly DraggableItem: boolean = false;
-    readonly SelectableItem: boolean = false;
-    readonly EnumerableItem: boolean = false;
-    readonly EdgeRowSeparatorItem: boolean = false;
-    readonly ItemActionsItem: boolean = false;
-
-    protected _$moreFontColorStyle: string;
-    protected _$moreCaption: string;
-
+export default class TreeNodeFooterItem extends TreeNodeExtraItem {
     readonly listInstanceName: string =  'controls-Tree__node-footer';
-
     readonly listElementName: string = 'item';
 
-    get node(): TreeItem<Model> {
-        return this.getNode();
-    }
-
-    getNode(): TreeItem<Model> {
-        return this.getParent();
-    }
-
     getTemplate(): TemplateFunction | string {
-        return this._$owner.getNodeFooterTemplate() || 'Controls/tree:NodeFooterTemplate';
+        return this._$owner.getNodeFooterTemplate() || super.getTemplate();
     }
 
     getItemClasses(): string {
-        return 'controls-ListView__itemV controls-Tree__nodeFooter';
+        return super.getItemClasses() + ' controls-Tree__nodeFooter';
     }
 
     getContentClasses(): string {
         return super.getContentClasses() + ' controls-Tree__itemContentTreeWrapper';
     }
 
-    getExpanderPaddingClasses(tmplExpanderSize?: string): string {
-        let classes = super.getExpanderPaddingClasses(tmplExpanderSize);
-
-        classes = classes.replace(
-           'controls-TreeGrid__row-expanderPadding',
-           'controls-TreeGrid__node-footer-expanderPadding'
-        );
-
-        return classes;
-    }
-
-    shouldDisplayVisibleFooter(content: TemplateFunction): boolean {
-        return this.hasMoreStorage('forward') || !!content;
-    }
-
-    getMoreFontColorStyle(): string {
-        return this._$moreFontColorStyle;
-    }
-
-    getMoreCaption(): string {
-        return this._$moreCaption;
-    }
-
-    setMoreFontColorStyle(moreFontColorStyle: string): void {
-        if (this._$moreFontColorStyle !== moreFontColorStyle) {
-            this._$moreFontColorStyle = moreFontColorStyle;
-            this._nextVersion();
-        }
-    }
-
-    protected _getLeftSpacingContentClasses(): string {
-        if (this._isDefaultRenderMultiSelect()) {
-            return ' controls-ListView__itemContent_withCheckboxes';
-        } else {
-            return ` controls-ListView__item-leftPadding_${this.getOwner().getLeftPadding().toLowerCase()}`;
-        }
+    getMoreClasses(): string {
+        return 'controls-TreeGrid__nodeFooterLoadMore';
     }
 }
 
 Object.assign(TreeNodeFooterItem.prototype, {
     '[Controls/tree:TreeNodeFooterItem]': true,
     _moduleName: 'Controls/tree:TreeNodeFooterItem',
-    _instancePrefix: 'tree-node-footer-item-',
-    _$moreFontColorStyle: null,
-    _$moreCaption: null
+    _instancePrefix: 'tree-node-footer-item-'
 });
