+++ Controls/_tree/render/NodeExtraItemTemplate.wml
@@ -30,7 +30,7 @@
                         loadMoreCaption="{{ item.getMoreCaption() }}"
                         linkFontColorStyle="{{ item.getMoreFontColorStyle() }}"
                         linkFontSize="xs"
-                        linkClass="controls-TreeGrid__nodeFooterLoadMore"/>
+                        linkClass="{{ item.getMoreClasses() }}"/>
         </ws:else>
     </div>
 </div>
